add Shoping CartPage, add Outlet in App, change routing to ShopingCartPage, change BackButtonSize, change footer
